+++ server/controllers/auth.ts
@@ -1,9 +1,8 @@
 import { Request, Response, NextFunction } from 'express';
 import { validationResult } from 'express-validator';
-import jwt from 'jsonwebtoken';
 import expressJwt from 'express-jwt';
-import User from '../models/user';
 import { IRequest } from 'server/interfaces/ExtendedRequest';
+import UsersService from '../services/AuthService';
 
 // eslint-disable-next-line consistent-return
 export const signup = async (req: Request, res: Response): Promise<any> => {
@@ -14,12 +13,16 @@ export const signup = async (req: Request, res: Response): Promise<any> => {
     });
   }
   try {
-    const user = new User(req.body);
-    const savedUser = await user.save();
+    const { name, email, id } = await UsersService.signup(
+      req.body.name,
+      req.body.lastName,
+      req.body.email,
+      req.body.password
+    );
     return res.json({
-      name: savedUser.name,
-      email: savedUser.email,
-      id: savedUser._id,
+      name,
+      email,
+      id,
     });
   } catch (error) {
     return res.status(400).json({
@@ -40,33 +43,15 @@ export const signin = async (req: Request, res: Response): Promise<any> => {
         message: errors.array()[0].msg,
       };
     }
-    const user = await User.findOne({ email: req.body.email });
-    if (!user) {
-      throw {
-        statusCode: 400,
-        message: 'USER not found!',
-      };
-    }
 
-    if (!user.authenticate(req.body.password)) {
-      throw {
-        statusCode: 401,
-        message: 'CREDENTIALS DO NOT MATCH!',
-      };
-    }
-
-    // TOKEN CREATED
-    const token = jwt.sign(
-      {
-        _id: user._id,
-      },
-      process.env.SECRET || ''
+    const { token, _id, name, email, role } = await UsersService.login(
+      req.body.email,
+      req.body.password
     );
 
     // PUTTING TOKEN INSIDE BROWSER OF USER
     res.cookie('token', token, { expires: new Date() });
 
-    const { _id, name, email, role } = user;
     return res.json({
       token,
       _id,
@@ -82,7 +67,7 @@ export const signin = async (req: Request, res: Response): Promise<any> => {
   }
 };
 
-export const signout = async (req: Request, res: Response): Promise<any> => {
+export const signout = async (_: Request, res: Response): Promise<any> => {
   res.clearCookie('token');
   return res.json({
     message: 'User signed out successfully',
